# kri8-main/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * kri8 API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkUserId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isPublic: boolean;
  themePreference: string;
  createdAt: string;
}

export interface UserPublic {
  id: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface UserUpdate {
  name?: string;
  username?: string;
  bio?: string;
  avatarUrl?: string;
  isPublic?: boolean;
  themePreference?: string;
}

export interface Idea {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  insight?: string | null;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  videoEditingNotes?: string | null;
  createdDate: string;
  /** @nullable */
  usedDate?: string | null;
  /** @nullable */
  customDate?: string | null;
  isUsed: boolean;
  /** @nullable */
  parentIdeaId?: number | null;
  branchCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface IdeaDetail {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  insight?: string | null;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  videoEditingNotes?: string | null;
  createdDate: string;
  /** @nullable */
  usedDate?: string | null;
  /** @nullable */
  customDate?: string | null;
  isUsed: boolean;
  /** @nullable */
  parentIdeaId?: number | null;
  branchCount: number;
  createdAt: string;
  updatedAt: string;
  branches: Idea[];
}

export interface IdeaInput {
  /** @minLength 1 */
  title: string;
  insight?: string;
  origin?: string;
  notes?: string;
  videoEditingNotes?: string;
  customDate?: string;
}

export interface IdeaUpdate {
  /** @minLength 1 */
  title?: string;
  insight?: string;
  origin?: string;
  notes?: string;
  videoEditingNotes?: string;
  customDate?: string;
  usedDate?: string;
  isUsed?: boolean;
  themePreference?: string;
}

export interface BranchInput {
  /** @minLength 1 */
  title: string;
  insight?: string;
  notes?: string;
  videoEditingNotes?: string;
}

export interface MarkUsedInput {
  usedDate?: string;
}

export interface IdeaStats {
  total: number;
  used: number;
  unused: number;
  withBranches: number;
  totalBranches: number;
  createdThisWeek: number;
  createdThisMonth: number;
}

export interface PublicProfile {
  user: UserPublic;
  ideas: IdeaDetail[];
}

export type FriendRequestStatus = typeof FriendRequestStatus[keyof typeof FriendRequestStatus];


export const FriendRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface FriendRequest {
  id: number;
  requester: UserPublic;
  addressee: UserPublic;
  status: FriendRequestStatus;
  createdAt: string;
}

export interface FriendsList {
  friends: UserPublic[];
  pendingReceived: FriendRequest[];
  pendingSent: FriendRequest[];
}

export type FriendRequestResponseStatus = typeof FriendRequestResponseStatus[keyof typeof FriendRequestResponseStatus];


export const FriendRequestResponseStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface FriendRequestResponse {
  status: FriendRequestResponseStatus;
}

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface Conversation {
  partner: UserPublic;
  lastMessage?: Message;
  unreadCount: number;
}

export interface TrendingTopic {
  id: string;
  name: string;
  category: string;
  growthPercent: number;
  volume: number;
  platform: string;
}

export interface TrendingHashtag {
  tag: string;
  platform: string;
  volume: number;
  growthPercent: number;
}

export interface ContentCategory {
  name: string;
  growthPercent: number;
  topContent: string[];
}

export type TrendDashboardProvider = typeof TrendDashboardProvider[keyof typeof TrendDashboardProvider];


export const TrendDashboardProvider = {
  mock: 'mock',
  youtube: 'youtube',
} as const;

export type TrendDashboardSource = typeof TrendDashboardSource[keyof typeof TrendDashboardSource];


export const TrendDashboardSource = {
  mock: 'mock',
  youtube: 'youtube',
} as const;

export type TrendDashboardMetricsQuality = typeof TrendDashboardMetricsQuality[keyof typeof TrendDashboardMetricsQuality];


export const TrendDashboardMetricsQuality = {
  fixture: 'fixture',
  estimated: 'estimated',
  measured: 'measured',
} as const;

export interface TrendDashboard {
  topics: TrendingTopic[];
  hashtags: TrendingHashtag[];
  categories: ContentCategory[];
  provider: TrendDashboardProvider;
  source: TrendDashboardSource;
  fetchedAt: string | null;
  isStatic: boolean;
  metricsQuality: TrendDashboardMetricsQuality;
}

export interface IdeaAnalysisRequest {
  title: string;
  notes?: string;
}

export interface IdeaAnalysisResult {
  relevanceScore: number | null;
  relatedTopics: TrendingTopic[];
  relatedHashtags: TrendingHashtag[];
  contentOpportunities: string[];
  formatAdaptations: string[];
  suggestedAngles: string[];
}

export interface InspirationRequest {
  title: string;
  notes?: string;
}

export type InspirationResultSource = typeof InspirationResultSource[keyof typeof InspirationResultSource];


export const InspirationResultSource = {
  openai: 'openai',
  template: 'template',
} as const;

export interface InspirationResult {
  relatedIdeas: string[];
  alternativeHooks: string[];
  titleSuggestions: string[];
  audienceQuestions: string[];
  source: InspirationResultSource;
}

export type SearchUsersParams = {
q: string;
};

export type ListIdeasParams = {
search?: string;
is_used?: boolean;
/**
 * @nullable
 */
parent_id?: number | null;
};

export type GetRecentIdeasParams = {
limit?: number;
};

export type GetCalendarIdeasParams = {
/**
 * Month in YYYY-MM format
 */
month: string;
};

export type GetMessagesParams = {
before?: number;
limit?: number;
};