# kri8-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: kri8 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trends/analyze:
    post:
      operationId: analyzeIdea
      tags: [trends]
      summary: Analyze an idea against current trends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaAnalysisRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaAnalysisResult"
  /trends/inspire:
    post:
      operationId: inspireIdea
      tags: [trends]
      summary: Generate content inspiration for an idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspirationRequest"
      responses:
        "200":
          description: Inspiration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspirationResult"
components:
  schemas:
...
    TrendingTopic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        growthPercent:
          type: number
        volume:
          type: number
        platform:
          type: string
      required: [id, name, category, growthPercent, volume, platform]
    TrendingHashtag:
      type: object
      properties:
        tag:
          type: string
        platform:
          type: string
        volume:
          type: number
        growthPercent:
          type: number
      required: [tag, platform, volume, growthPercent]
...
    IdeaAnalysisRequest:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
      required: [title]
    IdeaAnalysisResult:
      type: object
      properties:
        relevanceScore:
          type: number
          nullable: true
        relatedTopics:
          type: array
          items:
            $ref: "#/components/schemas/TrendingTopic"
        relatedHashtags:
          type: array
          items:
            $ref: "#/components/schemas/TrendingHashtag"
        contentOpportunities:
          type: array
          items:
            type: string
        formatAdaptations:
          type: array
          items:
            type: string
        suggestedAngles:
          type: array
          items:
            type: string
      required: [relevanceScore, relatedTopics, relatedHashtags, contentOpportunities, formatAdaptations, suggestedAngles]
    InspirationRequest:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
      required: [title]
    InspirationResult:
      type: object
      properties:
        relatedIdeas:
          type: array
          items:
            type: string
        alternativeHooks:
          type: array
          items:
            type: string
        titleSuggestions:
          type: array
          items:
            type: string
        audienceQuestions:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [openai, template]
      required: [relatedIdeas, alternativeHooks, titleSuggestions, audienceQuestions, source]