# kri8-main/lib/api-zod/src/generated/types/ideaAnalysisResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * kri8 API specification
 * OpenAPI spec version: 0.2.0
 */
import type { TrendingHashtag } from './trendingHashtag';
import type { TrendingTopic } from './trendingTopic';

export interface IdeaAnalysisResult {
  relevanceScore: number | null;
  relatedTopics: TrendingTopic[];
  relatedHashtags: TrendingHashtag[];
  contentOpportunities: string[];
  formatAdaptations: string[];
  suggestedAngles: string[];
}